study(step3): WinningChecker, LottoMatcher 의 역할과 로직을 Rank로 이관
