+++ src/main/kotlin/lotto/domain/model/Rank.kt
@@ -19,5 +19,30 @@ enum class Rank(val matchCount: Int, val prize: Int) {
                 rank.matchCount == matchCount
             } ?: NO_MATCH
         }
+
+        fun countMatchNumber(winningNumbers: WinningNumbers, lottoNumbers: List<Int>): Int {
+            return winningNumbers.numbers.sumOf { winningNumber ->
+                checkContainWinningNumber(winningNumber, lottoNumbers)
+            }
+        }
+
+        fun matchingWinner(matchCount: Int, matchBonus: Boolean): Rank {
+            check((Lotto.START_LOTTO_INDEX..Lotto.LAST_LOTTO_INDEX).contains(matchCount)) { "당첨 개수는 최대 6개 최소 0개 입니다" }
+            return Rank.safeValueOf(matchCount = matchCount, matchBonus = matchBonus)
+        }
+
+        fun win(winningNumbers: WinningNumbers, lottoNumbers: List<Int>, bonusNumber: Int = 0): Rank {
+            val matchCount: Int = Rank.countMatchNumber(winningNumbers, lottoNumbers)
+            val matchBonus: Boolean = matchBonus(matchCount, lottoNumbers, bonusNumber)
+            return Rank.matchingWinner(matchCount, matchBonus)
+        }
+
+        private fun matchBonus(matchCount: Int, lottoNumbers: List<Int>, bonusNumber: Int): Boolean {
+            return matchCount == 5 && lottoNumbers.contains(bonusNumber)
+        }
+
+        private fun checkContainWinningNumber(winningNumber: Int, lottoNumbers: List<Int>): Int {
+            return if (lottoNumbers.contains(winningNumber)) 1 else 0
+        }
     }
 }
